* ldaa: (23 commits)
  Connect Kyber Crypt functions to the TPM
  Add Kyber Encryption and Decryption functions
  Add final bug fixes to SignCommit2/3
  Add SignCommit2/3 Commands
  Fix bug in Commit 1 generation
  Fix bug in Kyber Key validation
  Pass size when serializing
  Fix bugs in LDAA implementation
  Add License to headers
  Add support for LDAA Sign Proceed command
  Clear LDAA internal state when TPM resets
  Fix bugs in create LDAA Key
  Fix command codes for LDAA
  TPM2_LDAA_SignCommit: Fill sign state once
  Add support for final LDAA command: Sign Proof
  Add Token Link command
  Update TPM to support LDAA
  Addresses half of the memory issues discussed
  Add commit 3 implementation for sign commit LDAA
  Add 2nd commit calculation in LDAA Sign Commit
  ...

+++ src/DuplicationCommands.c
@@ -71,7 +71,7 @@ TPM2_Duplicate(
 {
     TPM_RC                  result = TPM_RC_SUCCESS;
     TPMT_SENSITIVE          sensitive;
-    UINT16                  innerKeySize = 0; // encrypt key size for inner wrap
+    UINT32                  innerKeySize = 0; // encrypt key size for inner wrap
     OBJECT                  *object;
     OBJECT                  *newParent;
     TPM2B_DATA              data;
@@ -158,7 +158,7 @@ TPM2_Rewrap(
 {
     TPM_RC                  result = TPM_RC_SUCCESS;
     TPM2B_DATA              data;               // symmetric key
-    UINT16                  hashSize = 0;
+    UINT32                  hashSize = 0;
     TPM2B_PRIVATE           privateBlob;        // A temporary private blob
     // to transit between old
     // and new wrappers
@@ -187,7 +187,7 @@ TPM2_Rewrap(
 	    if(result != TPM_RC_SUCCESS)
 		return RcSafeAddToResult(result, RC_Rewrap_inDuplicate);
 	    // Copy unwrapped data to temporary variable, remove the integrity field
-	    hashSize = sizeof(UINT16) +
+	    hashSize = sizeof(UINT32) +
 		       CryptHashGetDigestSize(oldParent->publicArea.nameAlg);
 	    privateBlob.t.size = in->inDuplicate.t.size - hashSize;
 	    pAssert(privateBlob.t.size <= sizeof(privateBlob.t.buffer));
@@ -215,7 +215,7 @@ TPM2_Rewrap(
 	    if(result != TPM_RC_SUCCESS)
 		return result;
 	    // Copy temporary variable to output, reserve the space for integrity
-	    hashSize = sizeof(UINT16) +
+	    hashSize = sizeof(UINT32) +
 		       CryptHashGetDigestSize(newParent->publicArea.nameAlg);
 	    // Make sure that everything fits into the output buffer
 	    // Note: this is mostly only an issue if there was no outer wrapper on
@@ -262,7 +262,7 @@ TPM2_Import(
     TPMT_SENSITIVE           sensitive;
     TPM2B_NAME               name;
     TPMA_OBJECT              attributes;
-    UINT16                   innerKeySize = 0;       // encrypt key size for inner
+    UINT32                   innerKeySize = 0;       // encrypt key size for inner
     // wrapper
     // Input Validation
     // to save typing
